+++ packages/vite-plugin/src/client/es/loading-page-script.ts
