+++ packages/vite-plugin/src/client/html/loading-page.html
